Factor out components from Main to lighten its LoC
